requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.15.1
scikit-learn==1.6.1
joblib==1.3.1

# Machine Learning Models
xgboost==2.0.0
lightgbm==4.6.0
catboost==1.2.7

# Data Preprocessing & Optimization
dask==2025.2.0
dask[dataframe]==2025.2.0
numba==0.61.0

# Feature Importance
shap==0.46.0

# Google Cloud Storage (For GCP Integration)
google-cloud-storage==3.0.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
gcsfs==2025.2.0
# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==6.0.0

# Hyperparameter Tuning
optuna==4.2.1
optuna-integration==4.2.1

# General Utilities
tqdm==4.67.1
pyarrow==19.0.0
protobuf==4.25.6
requests==2.32.3
